requirements.txt
python-twitter
requests
ih
imgkit
